requirements.txt
aiogram==3.22.0
httpx==0.27.0
pydantic==2.11.7
pydantic-settings==2.10.1
qrcode[pil]==7.4.2
Pillow==10.4.0
python-dotenv==1.0.1
